notebook-teams/requirements.txt
aws-cdk.aws-sagemaker
aws-cdk.aws-efs
aws-cdk.cdk
aws-cdk.aws-iam
